+++ admin/modules/labels/delete.php
@@ -20,15 +20,23 @@
      */
 
 
-    $num = $db ->delete("labels", $id);
-    if($num > 0)
+    $is_product = $db->fetchOne("product", " labels_id = $id ");
+    debug($is_product);
+    if($is_product == NULL)
     {
-        $_SESSION['success'] = "Xóa thành công";
-        redirectAdmin("labels");
-    }
-    else
-    {
-        $_SESSION['error'] = "Xóa thất bại";
-        redirectAdmin("labels");
+         $num = $db ->delete("labels", $id);
+        if($num > 0)
+        {
+            $_SESSION['success'] = "Xóa thành công";
+            redirectAdmin("labels");
+        }
+        else
+        {
+            $_SESSION['error'] = "Xóa thất bại";
+            redirectAdmin("labels");
+        }   
+    }else {
+            $_SESSION['error'] = "Danh mục có sản phẩm bạn không thể xóa";
+            redirectAdmin("labels");
     }
 ?>
